requirements.txt
pcapy==0.11.1
dpkt==1.9.1
yara-python==3.6.3
IPy==0.83